silence cast warning
git-svn-id: https://aften.svn.sourceforge.net/svnroot/aften@267 ef0d8562-5c19-0410-972e-841db63a069c

+++ util/wavfilter.c
@@ -107,7 +107,7 @@ output_wav_data(FILE *ofp, FLOAT *samples, int ch, int n)
     int i;
 
     for(i=0; i<n*ch; i++) {
-        s16[0] = (samples[i] * 32767.0);
+        s16[0] = (int16_t)(samples[i] * 32767.0);
         write2le(*u16, ofp);
     }
 }
